Extracted types into models.go
